Optimize CSS and SCSS files for webpage layout
Made several adjustments in the custom.css and _travel-docs-template.scss files pertaining to the aesthetics and functionality of the website. This includes the removal of unnecessary comments, adding the "justify content" property for better alignment, and modifying the font-size parameters for better compatibility. These changes aim at improving web page display and layout.