Update CSS, HTML structure, and PHP for Private Template
This commit updates styles in "_travel-template.scss", "custom.css", and "_hero-video.scss" to support a new '.private-temp-hero-overlay' class. Also, the HTML structure in "private-template.php" and "functions.php" was updated for better recognition and operation with the new style. Furthermore, PHP adjustments were made in "private_image_field.php" to provide additional details and refactoring for better clarity and operation.