Update CSS for "call-to-action" block and other layout improvements
Several changes to the style.css have been made primarily aimed at improving the "call-to-action" block's visibility. The block now sports a grayish background (#e6e6e6) with black (#000) font color and is much more pronounced. Some minor tweaks were also made to the positioning of the "edit-link" and the bottom margin of the "content-page.entry-content". The changes anticipate improved layout and text readability on both the frontend and backend.